Add support for FILTER_NULL_ON_FAILURE flag
This works for all validates, but not sanitizes.

For now this lacks support for bit masks, e.g.:
`FILTER_FLAG_IPV4 | FILTER_NULL_ON_FAILURE`. The reason for this is
that there is also a `FILTER_FORCE_ARRAY` flag, which puts the
result in an array. This option, or bit masks, aren't supported either
right now. So adding support for bitmasks without `FILTER_FORCE_ARRAY`
would result in more false positives