Correct handling of GD outputs with no `filename` arg.
When the function is called as `imagepng($image)`, the output is returned,
and thus the expected return type is `string|false`.
The previous commits had the second parameter optional for both variants,
resulting in PHPStan inferring a return of `string|bool`.

Additionally, this parameter has been marked as optional for `imagexbm`,
in line with the PHP documentation.