Revert #265 "Add signatures for GD image*() returning strings"
The GD group of image output functions has the capability to not output to a
file, but the raw data is not returned by the function. Instead, the function
echos the data into the equivelant of `php://output`, and returns a boolean
for success or failure.

To quote the wording I previously misread on https://www.php.net/manual/en/function.imagepng.php:

> The path or an open stream resource [...] to save the file to.
> If not set or NULL, the raw image stream will be outputted directly.