Add failing test for #4657
While using the debugger I found that the problem occurs while resolving is_null($value).

This code:
https://github.com/phpstan/phpstan-src/blob/6d523028e399c15dc77aec3affd2ea97ff735925/src/Rules/Comparison/ImpossibleCheckTypeHelper.php#L177-L186
gets $sureTypes = [ Node\Expr\Variable  , Type\NullType ].

When treatPhpDocTypesAsCertain = false, it will call `getNativeType` on the variable. That resolves to `null`.

It's probably because the `$value = null;` declaration on the top.

It does not know that the value is also passed as reference to a callback.