type aliases: resolve type aliases lazily and in correct NameScope
This adds support for nested type aliases and helps preserve scope when importing aliases.

Aliases are resolved lazily because eager resolution would create a chicken-and-egg problem where aliases need to be resolved while they are already being resolved, thus not being resolved at all as a result of recursion protection.