Get closer to how type inference in TypeScript works
See https://javascript.xgqfrms.xyz/pdfs/TypeScript%20Language%20Specification.pdf
- section 3.11.7 Type Inference about union types

Otherwise, if T is a union or intersection type:
* First, inferences are made from S to each constituent type in T that isn't simply one of the
type parameters for which inferences are being made.
* If the first step produced no inferences then if T is a union type and exactly one
constituent type in T is simply a type parameter for which inferences are being made,
inferences are made from S to that type parameter.