Optimize inner loop of intersect union search
Array slice spends a good bit of time allocating memory for new arrays
even for small arrays. For larger computed unions this can be a
non-trivial amount of time. We can minimize the cost by not repeating
the slice for ever iteration of the inner loop since it doesn't depend
on the loop.