AnnotationsPropertiesClassReflectionExtension: Use bottom type for read/write-only properties
It is not meaningful to return the writable type from getReadableType() type for non-readable properties.
Let’s return a bottom type to make it clear that no value can be read from the property.
And similarly, for non-writable properties, getReadableType() will return bottom type
to emphasize that no value exists that can be written to the property.