Fully qualify the name of the `#[\Override]` attribute in error messages
This makes the attribute definition from the error message copy and paste safe
for namespaced classes.

As PHP attributes do not need to be backed by a class, using the non-qualified
`#[Override]` variant without a matching `use Override;` appears to look
correct at a quick glance, but will not actually do anything useful.