fix return type of sprintf with single %s format
In the return type extension of sprintf(), there was a shortcut to
assume the return type to be non-falsy-string when the format-string is
non-falsy string.

That's not correct though if the format string is just a single '%s'
placeholder (or a single '%x$s' placeholder). In that case, if the
requested argument is of stringy-type, then the return type is the type
of the stringy argument.

This commit extends the existing support for single-placeholder
templates that currently deals with numeric placeholders to also deal
with %s

this fixes #11201