Do not trigger NodeScopeResolver in data provider
When I'm debugging an issue with PHPStan, and want to go in deep into the
NodeScopeResolver, I'm always surprised that the first breakpoints are triggered
by this test. It's because work is done in the data provider that calls the NodeScopeResolver.

I think no work should be done in the PHPUnit data providers, because these are always executed.
Even when you specify a filter to only run a specific test.